+++ src/Components/Popover/index.tsx
@@ -1,13 +1,15 @@
+// Libraries
 import React, {Component} from 'react'
 
+// Components
 import {PopoverRoot, PopoverProps} from './Base/Popover'
 import {DismissButton} from '../Button/Composed/DismissButton'
 
 export class Popover extends Component<PopoverProps> {
   public static readonly displayName = 'Popover'
 
-  public static DismissButton = DismissButton
   public static Popover = PopoverRoot
+  public static DismissButton = DismissButton
 
   public render() {
     return <PopoverRoot {...this.props} />
